+++ src/ui/ExpansionEntrySetting.ts
